Rename some packages and classes of sample
