+++ sample/src/main/java/me/drakeet/multitype/sample/selectable/SquareViewBinder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package me.drakeet.multitype.sample.multiSelectable;
+package me.drakeet.multitype.sample.selectable;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
