+++ sample/src/main/res/layout/activity_more_apis_playground.xml
@@ -23,7 +23,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".moreApis.MoreApisPlayground">
+    tools:context=".more.MoreApisPlayground">
 
     <LinearLayout
         android:layout_width="match_parent"
