+++ sample/src/main/res/layout/activity_multi_selectable.xml
@@ -21,7 +21,7 @@
     android:id="@+id/activity_multi_grid"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="me.drakeet.multitype.sample.multiSelectable.MultiSelectableActivity">
+    tools:context="me.drakeet.multitype.sample.selectable.MultiSelectableActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/list"
